+++ asdf/tests/setup_package.py
@@ -7,4 +7,4 @@ from __future__ import absolute_import, division, unicode_literals, print_functi
 def get_package_data():  # pragma: no cover
     return {
         str(_PACKAGE_NAME_ + '.tests'):
-            ['coveragerc', 'data/*.yaml', 'data/*.fits']}
+            ['coveragerc', 'data/*.yaml', 'data/*.json', 'data/*.fits']}